style: SearchItem 컴포넌트 분리
