+++ src/app/(route)/(withLayout)/search/_components/SearchItemList.tsx
@@ -2,7 +2,6 @@
 
 import React, { useState, useEffect } from 'react'
 import Link from 'next/link'
-import { useRouter } from 'next/navigation'
 import Image from 'next/image'
 import { ItemType } from '@/app/_types/item.type'
 import useItemListData from '@/app/_hook/api/items/useItemListData'
@@ -10,56 +9,7 @@ import { cn } from '@/app/_utils/twMerge'
 import { useInView } from 'react-intersection-observer'
 import { SortOption } from '../_constants'
 import SortBox from './SortBox'
-
-export function Item({ item }: { item: ItemType }) {
-  const { name, price, id, image, createdAt, favoriteCount, reviewCount } =
-    item.itemSummary
-  const router = useRouter()
-
-  const handleItemClick = () => {
-    router.push(`/items/${id}`)
-  }
-
-  return (
-    <div className="flex w-[184px] flex-col gap-[7px] text-[14px]">
-      <Image
-        onClick={handleItemClick}
-        className="cursor-pointer rounded-[8px]"
-        src={image}
-        alt="item-image"
-        width={184}
-        height={184}
-      />
-      <div
-        onClick={handleItemClick}
-        className="line-clamp-2 h-[42px] cursor-pointer text-[#515151] hover:underline"
-      >
-        {name}
-      </div>
-      <strong>{price.toLocaleString()}원</strong>
-      <div className="flex gap-[20px] text-[13px] text-[#6F6F6F]">
-        <div className="flex cursor-pointer">
-          <Image
-            src="/image/icon/icon-save.svg"
-            alt="save"
-            width={18}
-            height={18}
-          />
-          <div>{favoriteCount}</div>
-        </div>
-        <div className="flex cursor-pointer">
-          <Image
-            src="/image/icon/icon-review.svg"
-            alt="review"
-            width={18}
-            height={18}
-          />
-          <div>{reviewCount}</div>
-        </div>
-      </div>
-    </div>
-  )
-}
+import { SearchItem } from './SearchItem'
 
 export default function SearchItemList({ keyword }: { keyword: string }) {
   const [sortOption, setSortOption] = useState(SortOption[2])
@@ -115,7 +65,7 @@ export default function SearchItemList({ keyword }: { keyword: string }) {
       </div>
       <div className="grid grid-cols-[repeat(auto-fill,184px)] gap-x-[19px] gap-y-[25px]">
         {itemList.map((item: ItemType) => {
-          return <Item item={item} key={item.cursorId} />
+          return <SearchItem item={item} key={item.cursorId} />
         })}
       </div>
       {isFetchingNextPage ? <div>Loading..</div> : <div ref={ref} />}
